requirements.txt
numpy~=1.24.2
nrlmsise00~=0.1.2
scipy~=1.11.4
matplotlib~=3.7.1
mpld3~=0.5.10
tqdm~=4.65.0
plotly~=5.18.0
numba~=0.58.0
pyproj~=3.6.1
iri2016~=1.11.1
Pillow~=9.4.0